move resource manager, add local run check
